requirements.txt
fastapi[all]
uvicorn[standard]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
cryptography
python-jose[cryptography]
passlib